requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1  # Aligns with SQLAlchemy 2.0.41
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.1
email-validator==2.2.0
python-dotenv==1.1.0
itsdangerous==2.2.0
Flask-Mail==0.10.0
boto3==1.38.39
# For video processing (subprocess calls to ffmpeg)
moviepy==2.2.1
# For async tasks (optional, comment out if not used)
# celery==5.5.3
# redis==5.1.1
Werkzeug==3.1.3
# For production deployment (optional)
# gunicorn==23.0.0
# psycopg2-binary==2.9.10

# Dev tools
black
flake8
pytest